requirements.txt
alembic==1.0.6
autopep8==1.4.3
bcrypt==3.1.6
blinker==1.4
cffi==1.11.5
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.9.4
flask-wysiwyg==0.1
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mysqlclient==1.4.2.post1
Pillow==5.4.1
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.16
visitor==0.1.3
Werkzeug==0.14.1
WTForms==1.0.5
